The carto 3Bs map with toggleable layers
This was an attempt to 1) make the layer toggle more visible (the built in one in Carto is kind of cryptic) and 2) restrict it so that people can only view one layer at a time.  Additional tasks to be completed include improving the toggle buttons and seeing if there's a way to increase the size of the legend.